+++ src/events.h
@@ -2,6 +2,7 @@
 
 enum CustomEvent {
   about_dialog_requested,
+  change_unit_system,
   model_updated,
   new_location_requested,
   refresh_requested,